backend/content_service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyMySQL==1.1.0
cryptography==41.0.7
celery==5.3.4
redis==5.0.1
python-decouple==3.8
requests==2.31.0
gunicorn==21.2.0
django-filter==23.3
